Comma for score
Changed all score in code to be int.
made a function that adds a comma to every thousand points in the score counter of the HUD.

+++ Source/FrogGame/FrogGameCharacter.h
@@ -82,7 +82,7 @@ public:
 
 	/** Accessor Function for Current Score */
 	UFUNCTION(BlueprintCallable, Category = "Character | Score")
-	float GetCurrentScore() const
+	int GetCurrentScore() const
 	{
 		return CurrentScore;
 	}
@@ -91,7 +91,7 @@ public:
 	* @Param Score This is the amount to increase the players score by. This should only be positive!
 	*/
 	UFUNCTION(BlueprintCallable, Category = "Character | Score")
-	void UpdateCurrentScore(const float Score)
+	void UpdateCurrentScore(const int Score)
 	{
 		CurrentScore = CurrentScore + Score;
 	}
@@ -158,7 +158,7 @@ public:
 
 	/** The Players current score */
 	UPROPERTY(EditAnywhere, SaveGame, Category = "Character | Score")
-	float CurrentScore;
+	int CurrentScore;
 	// How close object has to be to be eaten (destroyed).
 	UPROPERTY(EditDefaultsOnly, BlueprintReadOnly, Category = "Character | Whirlwind")
 	float EatDistance{150.f};
